ml/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0
